[1234] fix review comments + add unit tests
