+++ src/models/user-token-profile.model.ts
@@ -1,16 +1,6 @@
 import { UserRole } from './user.db.model';
 
-export interface JWTTokenPayload {
-    id: string;
-    username: string;
-    role: UserRole;
-    firstname: string;
-    lastname: string;
-    exp: number;
-    class_id?: string;
-}
-
-export interface UserTokenProfile {
+export class UserTokenProfile {
     id: string;
     username: string;
     role: UserRole;
